Refactor GetProxyHttpClient().
