+++ conf/app.conf
@@ -28,4 +28,4 @@ mailUser = ""
 mailPass = ""
 mailHost = ""
 mailPort = ""
-httpproxy = ""
+httpProxy = ""
